src/main/java/pt/ipp/isep/dei/project/dto/LocalDTO.java
package pt.ipp.isep.dei.project.dto;


public class LocalDTO {

    private double latitude;
    private double longitude;
    private double altitude;
    private long id;

    public double getLatitude() {
        return latitude;
    }

    public void setLatitude(double latitude) {
        this.latitude = latitude;
    }

    public double getLongitude() {
        return longitude;
    }

    public void setLongitude(double longitude) {
        this.longitude = longitude;
    }

    public double getAltitude() {
        return altitude;
    }

    public void setAltitude(double altitude) {
        this.altitude = altitude;
    }

    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }
}